Add files via upload
The global COVID-19 pandemic has created exceptional difficulties for
college and university students. With educational institutions shutting
down and an abrupt shift to online learning, students are facing new
challenges. Despite the advantages of e-learning, developing nations
like Pakistan are experiencing major hurdles due to a lack of internet
access and financial constraints. As a result, students' academic
performance is being negatively impacted. Researchers are using
sentiment analysis of social media platforms such as Facebook, Twitter,
and YouTube to gain insight into student experiences during this
challenging time. This study seeks to create a machine learning-based
sentiment classification system that can determine the sentiment of
higher education students in Pakistan towards e-learning classes during
the COVID-19 pandemic. The primary objective of this research is to
identify and categorize the challenges and problems that university
students face in their e-learning classes. The research employed MultiNominal NB and Gaussian NB classifiers to examine the sentiment of
online reviews and comments of students. We achieve 99% accuracy
using Gaussian NB and 98% accuracy using Multinomial NB classifiers.
According to the study's results, the COVID-19 epidemic has had a
significant effect on the education system, with the vast majority of
feedback being negative (70%). Only a small percentage of feedback
was classified as neutral (29.9%) or positive (4%). This finding underscores
the urgent need for better solutions for online learning during and after
the pandemic. The study underscores the importance of addressing
these unique challenges faced by students in developing countries
during the pandemic. There is a pressing need for innovative solutions
that can ensure that higher education remains effective and accessible
in these challenging times. By utilizing sentiment analysis and machine
learning, researchers can gain a better understanding of the
experiences of students during this crisis. This understanding can inform
targeted interventions and solutions that can improve academic
performance and well-being for students in developing countries like
Pakistan. The findings of this research provide important insights and
recommendations for policymakers, educators, and stakeholders in the
higher education sector.